automation/telegram_bot/requirements.txt
# Telegram Bot Dependencies
python-telegram-bot==20.7
python-telegram-bot[webhooks]==20.7

# Database and Storage
asyncpg==0.29.0
redis[hiredis]==5.0.1
sqlalchemy==2.0.23

# Payment Processing
stripe==7.8.0

# Web Framework (for webhooks)
fastapi==0.104.1
uvicorn[standard]==0.24.0
aiohttp==3.9.1
httpx==0.25.2
uiautomator2==2.16.23

# Data Handling
pydantic==1.10.13
python-multipart==0.0.6

# Utilities
python-dotenv==1.0.0
schedule==1.2.0
asyncio-mqtt==0.13.0

# Monitoring and Logging
structlog==23.2.0
sentry-sdk==1.38.0
hvac==1.1.1

# Development and Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0